10764 : Give specific login message to user whose account has been deactiviated

+++ api/src/org/labkey/api/data/Table.java
@@ -119,7 +119,7 @@ public class Table
             Statement statement = conn.createStatement(scrollable ? ResultSet.TYPE_SCROLL_INSENSITIVE : ResultSet.TYPE_FORWARD_ONLY, ResultSet.CONCUR_READ_ONLY);
             if (null != statementRowCount)
             {
-                statement.setMaxRows(statementRowCount.intValue() == NO_ROWS ? 0 : statementRowCount.intValue());
+                statement.setMaxRows(statementRowCount == NO_ROWS ? 0 : statementRowCount);
             }
             if (asyncRequest != null)
             {
@@ -132,7 +132,7 @@ public class Table
             PreparedStatement stmt = conn.prepareStatement(sql, scrollable ? ResultSet.TYPE_SCROLL_INSENSITIVE : ResultSet.TYPE_FORWARD_ONLY, ResultSet.CONCUR_READ_ONLY);
             if (null != statementRowCount)
             {
-                stmt.setMaxRows(statementRowCount.intValue());
+                stmt.setMaxRows(statementRowCount);
             }
             if (asyncRequest != null)
             {
